+++ chat-widget/src/common/storage/default/defaultClientDataStoreProvider.ts
@@ -1,20 +1,31 @@
-import { inMemoryDataStore } from "./defaultInMemoryDataStore";
-import { IContextDataStore } from "../../interfaces/IContextDataStore";
-import { TelemetryHelper } from "../../telemetry/TelemetryHelper";
+/* eslint-disable @typescript-eslint/no-explicit-any */
+
 import { LogLevel, TelemetryEvent } from "../../telemetry/TelemetryConstants";
 
-export const defaultClientDataStoreProvider = (): IContextDataStore => {
+import { IContextDataStore } from "../../interfaces/IContextDataStore";
+import { TelemetryHelper } from "../../telemetry/TelemetryHelper";
+import { inMemoryDataStore } from "./defaultInMemoryDataStore";
+
+export const defaultClientDataStoreProvider = (cacheTtlinMins = 0): IContextDataStore => {
+    let ttlInMs = 0;
     const isCookieAllowed = () => {
         try {
             localStorage;
             sessionStorage;
             return true;
         } catch (error) {
-            console.error("Third party cookie blocked");
+            if (!(window as any).TPCWarningShown) {
+                console.warn("Third party cookies blocked.");
+                (window as any).TPCWarningShown = true;
+            }
             return false;
         }
     };
-    const TtlInMs = 15 * 60 * 1000; // 15 mins
+
+    if (ttlInMs == 0) {
+        ttlInMs = cacheTtlinMins * 60 * 1000;
+    }
+    
     const dataStoreProvider = {
         // eslint-disable-next-line @typescript-eslint/no-explicit-any
         setData: (key: any, data: any, type: any) => {
@@ -24,7 +35,7 @@ export const defaultClientDataStoreProvider = (): IContextDataStore => {
                         const now = new Date();
                         const item = {
                             data: data,
-                            expiry: now.getTime() + TtlInMs,
+                            expiry: now.getTime() + ttlInMs,
                         };
                         const strItem = JSON.stringify(item);